Added wallet db modules
